+++ CHANGELOG.md
@@ -16,6 +16,8 @@
 * Fix: misbehavour related to 'NOT' and '=' predecession
 * Fix: alasql running from cordova on iOS
 * Fix: SELECT can use functions from GROUP BY list
+* Fix: Bug in NUMERIC type conversion
+* Added: functions CEIL/CEILING and FLOOR
 
 ## Version 0.2.0 "Athens" (13.07.2015)
 The purpose of this release were hard work on: